feat: sync selected Vinh features

+++ backend/src/main/java/com/cinebook/backend/modules/dashboard/dto/KpiResponse.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class KpiResponse {
     private Integer totalRevenue;
-    private Long totalTickets;
+    private Integer totalTickets;
     private Long newUsers;
 }
